Move all PostgresConnection code into one file
