--- Sources/PostgresNIO/Connection/PostgresConnection+Authenticate.swift
@@ -1,24 +1,0 @@
-import NIOCore
-import Logging
-
-extension PostgresConnection {
-    public func authenticate(
-        username: String,
-        database: String? = nil,
-        password: String? = nil,
-        logger: Logger = .init(label: "codes.vapor.postgres")
-    ) -> EventLoopFuture<Void> {
-        let authContext = AuthContext(
-            username: username,
-            password: password,
-            database: database)
-        let outgoing = PSQLOutgoingEvent.authenticate(authContext)
-        self.underlying.channel.triggerUserOutboundEvent(outgoing, promise: nil)
-
-        return self.underlying.channel.pipeline.handler(type: PSQLEventsHandler.self).flatMap { handler in
-            handler.authenticateFuture
-        }.flatMapErrorThrowing { error in
-            throw error.asAppropriatePostgresError
-        }
-    }
-}
